requirements.txt
requests
pyyaml
pandas
pytz
python-dotenv
matplotlib
japanize-matplotlib
plotly